requirements.txt
appium-python-client>=3.0.0
selenium>=4.8.0
pandas>=2.0.0
openpyxl>=3.0.0
